move and rename pool Metadata, Metrics and Performance inside the Jörmungandr package
The cardano-node implementation will actually be drastically different, to instead of trying to share any sort of implementation here,
we'll instead revise this implementation completely. This follows the original plan and explains why the 'Pool' modules were kept
isolated from the rest: to be easy to remove and replace.
