+++ lib/core/cardano-wallet-core.cabal
@@ -113,10 +113,6 @@ library
       Cardano.Pool.DB.Model
       Cardano.Pool.DB.Sqlite
       Cardano.Pool.DB.Sqlite.TH
-      Cardano.Pool.Metrics
-      Cardano.Pool.Metadata
-      Cardano.Pool.Performance
-      Cardano.Pool.Ranking
       Cardano.Wallet
       Cardano.Wallet.Api
       Cardano.Wallet.Api.Client
@@ -256,11 +252,6 @@ test-suite unit
       Cardano.Pool.DB.Arbitrary
       Cardano.Pool.DB.Properties
       Cardano.Pool.DB.SqliteSpec
-      Cardano.Pool.MetadataSpec
-      Cardano.Pool.MetricsSpec
-      Cardano.Wallet.Gen
-      Cardano.Pool.PerformanceSpec
-      Cardano.Pool.RankingSpec
       Cardano.Wallet.Api.Malformed
       Cardano.Wallet.Api.Server.TlsSpec
       Cardano.Wallet.Api.ServerSpec
@@ -273,6 +264,7 @@ test-suite unit
       Cardano.Wallet.DB.Sqlite.TypesSpec
       Cardano.Wallet.DB.StateMachine
       Cardano.Wallet.DummyTarget.Primitive.Types
+      Cardano.Wallet.Gen
       Cardano.Wallet.Network.BlockHeadersSpec
       Cardano.Wallet.Network.PortsSpec
       Cardano.Wallet.NetworkSpec
